remove build.pl again; replace by all new `b` command
`b` is much better in what it does than build.pl: It does not install
globally anymore, but automatically creates build prefixes ("buckets")
in the home directory. The tree-bucket relation can survive "rm -r $B".

Building a source tree is as simple as typing `b` on the CLI and
entering a bucket name. Everything else is automatic from there.
Building and installing is `b && b i`, or
`b configure && b make && b install` if you like to be explicit.

`b` is built on a OOP-style Perl module Build.pm which I want to use to
integrate kdesrc-build with my workflow.